# backend/openapi/openapi.json
{
  "components": {
    "schemas": {
...
      "Court": {
        "properties": {
          "created": {
            "format": "date-time",
            "title": "Created",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "tournament_id": {
            "title": "Tournament Id",
            "type": "integer"
          }
        },
        "required": [
          "name",
          "created",
          "tournament_id",
          "id"
        ],
        "title": "Court",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MatchScoreTrackingBody": {
        "properties": {
          "stage_item_input1_score": {
            "default": 0,
            "title": "Stage Item Input1 Score",
            "type": "integer"
          },
          "stage_item_input2_score": {
            "default": 0,
            "title": "Stage Item Input2 Score",
            "type": "integer"
          },
          "state": {
            "$ref": "#/components/schemas/MatchState",
            "default": "NOT_STARTED"
          }
        },
        "required": [
          "stage_item_input1_score",
          "stage_item_input2_score",
          "state"
        ],
        "title": "MatchScoreTrackingBody",
        "type": "object"
      },
      "MatchState": {
        "enum": [
          "NOT_STARTED",
          "IN_PROGRESS",
          "COMPLETED"
        ],
        "title": "MatchState",
        "type": "string"
      },
      "MatchWithDetails": {
        "description": "MatchWithDetails has zero or one defined stage item inputs, but not both.",
        "properties": {
          "completed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "court": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Court"
              },
              {
                "type": "null"
              }
            ]
          },
          "court_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Court Id"
          },
          "created": {
            "format": "date-time",
            "title": "Created",
            "type": "string"
          },
          "custom_duration_minutes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Custom Duration Minutes"
          },
          "custom_margin_minutes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Custom Margin Minutes"
          },
          "duration_minutes": {
            "title": "Duration Minutes",
            "type": "integer"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "margin_minutes": {
            "title": "Margin Minutes",
            "type": "integer"
          },
          "position_in_schedule": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Position In Schedule"
          },
          "round_id": {
            "title": "Round Id",
            "type": "integer"
          },
          "stage_item_input1": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/StageItemInputTentative"
              },
              {
                "$ref": "#/components/schemas/StageItemInputFinal"
              },
              {
                "$ref": "#/components/schemas/StageItemInputEmpty"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stage Item Input1"
          },
          "stage_item_input1_conflict": {
            "title": "Stage Item Input1 Conflict",
            "type": "boolean"
          },
          "stage_item_input1_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stage Item Input1 Id"
          },
          "stage_item_input1_score": {
            "title": "Stage Item Input1 Score",
            "type": "integer"
          },
          "stage_item_input1_winner_from_match_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stage Item Input1 Winner From Match Id"
          },
          "stage_item_input2": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/StageItemInputTentative"
              },
              {
                "$ref": "#/components/schemas/StageItemInputFinal"
              },
              {
                "$ref": "#/components/schemas/StageItemInputEmpty"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stage Item Input2"
          },
          "stage_item_input2_conflict": {
            "title": "Stage Item Input2 Conflict",
            "type": "boolean"
          },
          "stage_item_input2_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stage Item Input2 Id"
          },
          "stage_item_input2_score": {
            "title": "Stage Item Input2 Score",
            "type": "integer"
          },
          "stage_item_input2_winner_from_match_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stage Item Input2 Winner From Match Id"
          },
          "start_time": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Time"
          },
          "state": {
            "$ref": "#/components/schemas/MatchState",
            "default": "NOT_STARTED"
          }
        },
        "required": [
          "created",
          "start_time",
          "duration_minutes",
          "margin_minutes",
          "custom_duration_minutes",
          "custom_margin_minutes",
          "position_in_schedule",
          "round_id",
          "stage_item_input1_score",
          "stage_item_input2_score",
          "court_id",
          "stage_item_input1_conflict",
          "stage_item_input2_conflict",
          "state",
          "completed_at",
          "stage_item_input1_id",
          "stage_item_input2_id",
          "stage_item_input1_winner_from_match_id",
          "stage_item_input2_winner_from_match_id",
          "id",
          "stage_item_input1",
          "stage_item_input2",
          "court"
        ],
        "title": "MatchWithDetails",
        "type": "object"
      },
...
      "ScoreTrackingInfo": {
        "properties": {
          "matches": {
            "items": {
              "$ref": "#/components/schemas/MatchWithDetails"
            },
            "title": "Matches",
            "type": "array"
          },
          "tournament_id": {
            "title": "Tournament Id",
            "type": "integer"
          },
          "tournament_name": {
            "title": "Tournament Name",
            "type": "string"
          }
        },
        "required": [
          "tournament_id",
          "tournament_name",
          "matches"
        ],
        "title": "ScoreTrackingInfo",
        "type": "object"
      },
      "ScoreTrackingInfoResponse": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/ScoreTrackingInfo"
          }
        },
        "required": [
          "data"
        ],
        "title": "ScoreTrackingInfoResponse",
        "type": "object"
      },
      "ScoreTrackingMatchResponse": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/MatchWithDetails"
          }
        },
        "required": [
          "data"
        ],
        "title": "ScoreTrackingMatchResponse",
        "type": "object"
      },
...
      "StageItemInputEmpty": {
        "properties": {
          "draws": {
            "default": 0,
            "title": "Draws",
            "type": "integer"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "losses": {
            "default": 0,
            "title": "Losses",
            "type": "integer"
          },
          "points": {
            "default": "0.0",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Points",
            "type": "string"
          },
          "slot": {
            "title": "Slot",
            "type": "integer"
          },
          "stage_item_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stage Item Id"
          },
          "team_id": {
            "title": "Team Id",
            "type": "null"
          },
          "tournament_id": {
            "title": "Tournament Id",
            "type": "integer"
          },
          "winner_from_stage_item_id": {
            "title": "Winner From Stage Item Id",
            "type": "null"
          },
          "winner_position": {
            "title": "Winner Position",
            "type": "null"
          },
          "wins": {
            "default": 0,
            "title": "Wins",
            "type": "integer"
          }
        },
        "required": [
          "points",
          "wins",
          "draws",
          "losses",
          "id",
          "slot",
          "tournament_id",
          "stage_item_id",
          "team_id",
          "winner_from_stage_item_id",
          "winner_position"
        ],
        "title": "StageItemInputEmpty",
        "type": "object"
      },
      "StageItemInputFinal": {
        "properties": {
          "draws": {
            "default": 0,
            "title": "Draws",
            "type": "integer"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "losses": {
            "default": 0,
            "title": "Losses",
            "type": "integer"
          },
          "points": {
            "default": "0.0",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Points",
            "type": "string"
          },
          "slot": {
            "title": "Slot",
            "type": "integer"
          },
          "stage_item_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stage Item Id"
          },
          "team": {
            "$ref": "#/components/schemas/Team"
          },
          "team_id": {
            "title": "Team Id",
            "type": "integer"
          },
          "tournament_id": {
            "title": "Tournament Id",
            "type": "integer"
          },
          "winner_from_stage_item_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Winner From Stage Item Id"
          },
          "winner_position": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Winner Position"
          },
          "wins": {
            "default": 0,
            "title": "Wins",
            "type": "integer"
          }
        },
        "required": [
          "points",
          "wins",
          "draws",
          "losses",
          "id",
          "slot",
          "tournament_id",
          "stage_item_id",
          "team_id",
          "winner_from_stage_item_id",
          "winner_position",
          "team"
        ],
        "title": "StageItemInputFinal",
        "type": "object"
      },
...
      "StageItemInputTentative": {
        "properties": {
          "draws": {
            "default": 0,
            "title": "Draws",
            "type": "integer"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "losses": {
            "default": 0,
            "title": "Losses",
            "type": "integer"
          },
          "points": {
            "default": "0.0",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Points",
            "type": "string"
          },
          "slot": {
            "title": "Slot",
            "type": "integer"
          },
          "stage_item_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stage Item Id"
          },
          "team_id": {
            "title": "Team Id",
            "type": "null"
          },
          "tournament_id": {
            "title": "Tournament Id",
            "type": "integer"
          },
          "winner_from_stage_item_id": {
            "title": "Winner From Stage Item Id",
            "type": "integer"
          },
          "winner_position": {
            "minimum": 1.0,
            "title": "Winner Position",
            "type": "integer"
          },
          "wins": {
            "default": 0,
            "title": "Wins",
            "type": "integer"
          }
        },
        "required": [
          "points",
          "wins",
          "draws",
          "losses",
          "id",
          "slot",
          "tournament_id",
          "stage_item_id",
          "team_id",
          "winner_from_stage_item_id",
          "winner_position"
        ],
        "title": "StageItemInputTentative",
        "type": "object"
      },
...
      "Team": {
        "properties": {
          "active": {
            "title": "Active",
            "type": "boolean"
          },
          "created": {
            "format": "date-time",
            "title": "Created",
            "type": "string"
          },
          "draws": {
            "default": 0,
            "title": "Draws",
            "type": "integer"
          },
          "elo_score": {
            "default": "1200",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Elo Score",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "logo_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Logo Path"
          },
          "losses": {
            "default": 0,
            "title": "Losses",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "swiss_score": {
            "default": "0.0",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Swiss Score",
            "type": "string"
          },
          "tournament_id": {
            "title": "Tournament Id",
            "type": "integer"
          },
          "wins": {
            "default": 0,
            "title": "Wins",
            "type": "integer"
          }
        },
        "required": [
          "created",
          "name",
          "tournament_id",
          "active",
          "elo_score",
          "swiss_score",
          "wins",
          "draws",
          "losses",
          "logo_path",
          "id"
        ],
        "title": "Team",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "token"
          }
        },
        "type": "oauth2"
      }
    }
  },
  "info": {
    "description": "\n### Description\nThis API allows you to do everything the frontend of [Bracket](https://github.com/evroon/bracket)\nallows you to do (the frontend uses this API as well).\n\nFore more information, see the [documentation](https://docs.bracketapp.nl).\n\n### Table of Contents\n*(links only work for [ReDoc](https://api.bracketapp.nl/redoc), not for Swagger UI)*\n\n- [Auth](#tag/Auth)\n\n- [Clubs](#tag/Clubs)\n\n- [Courts](#tag/Courts)\n\n- [Internals](#tag/Internals)\n\n- [Matches](#tag/Matches)\n\n- [Players](#tag/Players)\n\n- [Rankings](#tag/Rankings)\n\n- [Rounds](#tag/Rounds)\n\n- [Score Tracking](#tag/Score-Tracking)\n\n- [Signup](#tag/Signup)\n\n- [Stage Items](#tag/Stage-Items)\n\n- [Stage Item Inputs](#tag/Stage-Item-Inputs)\n\n- [Stages](#tag/Stages)\n\n- [Teams](#tag/Teams)\n\n- [Tournaments](#tag/Tournaments)\n\n- [Users](#tag/Users)\n\n### Links\nGitHub: <https://github.com/evroon/bracket>\n\nDocs: <https://docs.bracketapp.nl>\n\nDemo: <https://www.bracketapp.nl/demo>\n\nAPI docs (Redoc): <https://api.bracketapp.nl/redoc>\n\nAPI docs (Swagger UI): <https://api.bracketapp.nl/docs>\n",
    "license": {
      "name": "AGPL-3.0",
      "url": "https://www.gnu.org/licenses/agpl-3.0.en.html"
    },
    "summary": "API for Bracket, an open source tournament system.",
    "title": "Bracket API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/score-tracking/{score_tracking_token}": {
      "get": {
        "operationId": "get_score_tracking_info_score_tracking__score_tracking_token__get",
        "parameters": [
          {
            "in": "path",
            "name": "score_tracking_token",
            "required": true,
            "schema": {
              "title": "Score Tracking Token",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScoreTrackingInfoResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Score Tracking Info",
        "tags": [
          "Score Tracking"
        ]
      }
    },
    "/score-tracking/{score_tracking_token}/matches/{match_id}": {
      "get": {
        "operationId": "get_score_tracking_match_score_tracking__score_tracking_token__matches__match_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "match_id",
            "required": true,
            "schema": {
              "title": "Match Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "score_tracking_token",
            "required": true,
            "schema": {
              "title": "Score Tracking Token",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScoreTrackingMatchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Score Tracking Match",
        "tags": [
          "Score Tracking"
        ]
      },
      "put": {
        "operationId": "update_score_tracking_match_score_tracking__score_tracking_token__matches__match_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "match_id",
            "required": true,
            "schema": {
              "title": "Match Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "score_tracking_token",
            "required": true,
            "schema": {
              "title": "Score Tracking Token",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MatchScoreTrackingBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScoreTrackingMatchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Score Tracking Match",
        "tags": [
          "Score Tracking"
        ]
      }
    },
...
  }
}